changing adding styles and setting entry point
