+++ webpack.config.js
@@ -1,6 +1,7 @@
 	const path = require('path');
 
 	const HtmlWebpackPlugin = require('html-webpack-plugin');
+	const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
 	module.exports = (env) => {
 		return {
@@ -13,14 +14,17 @@
 			},
 			plugins: [
 				new HtmlWebpackPlugin({
-					template: path.resolve(__dirname, 'src', 'index.html')
+					template: path.resolve(__dirname, 'src', 'index.html'),
+				}),
+				new MiniCssExtractPlugin({
+					filename: 'style.css',
 				}),
 			],
 			module: {
 				rules: [
 					{
 						test: /\.css$/i,
-						use: ["style-loader", "css-loader"],
+						use: [MiniCssExtractPlugin.loader, "css-loader"],
 					},
 				],
 			},
